Fix jasmine specs
All the specs work now.  There were 2 issues.

1. Because we were overriding the route function instead of updating
the routes hash, the routes hash wasn't being updated.  I've added code
to the route function to put the new route in the routes hash.  Some of
the failures were due to the test expecting the prefixed version of the
routes to be in the routes hash.
2. There was a genuine bug wherein if you added the "" default route,
the route that ended up getting added was "prefix/" instead of the
correct "prefix".  Fixed.